fix api key generation when non admin
but this will need to be better moved around

+++ app/controllers/UsersController.php
@@ -20,12 +20,25 @@ $redirect = true;
 try {
     require_once '../../app/init.inc.php';
 
-    if (!$_SESSION['is_admin']) {
-        throw new Exception('Non admin user tried to access admin panel.');
-    }
-
     $FormKey = new FormKey();
     $Users = new Users(null, new Config);
+
+    // (RE)GENERATE AN API KEY (from profile)
+    if (isset($_POST['generateApiKey'])) {
+        $redirect = false;
+        $Users->setId($_SESSION['userid']);
+        if ($Users->generateApiKey()) {
+            echo json_encode(array(
+                'res' => true,
+                'msg' => _('Saved')
+            ));
+        } else {
+            echo json_encode(array(
+                'res' => false,
+                'msg' => Tools::error()
+            ));
+        }
+    }
 
     $tab = 1;
     $location = '../../admin.php?tab=' . $tab;
@@ -33,6 +46,10 @@ try {
     // VALIDATE
     if (isset($_POST['usersValidate'])) {
         $tab = 2;
+        if (!$_SESSION['is_admin']) {
+            throw new Exception('Non admin user tried to access admin panel.');
+        }
+
         // loop the array
         foreach ($_POST['usersValidateIdArr'] as $userid) {
             $_SESSION['ok'][] = $Users->validate($userid);
@@ -42,6 +59,9 @@ try {
     // UPDATE USERS
     if (isset($_POST['usersUpdate'])) {
         $tab = 2;
+        if (!$_SESSION['is_admin']) {
+            throw new Exception('Non admin user tried to access admin panel.');
+        }
         if (isset($_POST['fromSysconfig'])) {
             $location = "../../sysconfig.php?tab=$tab";
         } else {
@@ -59,6 +79,9 @@ try {
         && isset($_POST['usersDestroy'])) {
 
         $tab = 2;
+        if (!$_SESSION['is_admin']) {
+            throw new Exception('Non admin user tried to access admin panel.');
+        }
 
         if ($Users->destroy(
             $_POST['usersDestroyEmail'],
@@ -67,24 +90,6 @@ try {
             $_SESSION['ok'][] = _('Everything was purged successfully.');
         }
     }
-
-    // (RE)GENERATE AN API KEY
-    if (isset($_POST['generateApiKey'])) {
-        $redirect = false;
-        $Users->setId($_SESSION['userid']);
-        if ($Users->generateApiKey()) {
-            echo json_encode(array(
-                'res' => true,
-                'msg' => _('Saved')
-            ));
-        } else {
-            echo json_encode(array(
-                'res' => false,
-                'msg' => Tools::error()
-            ));
-        }
-    }
-
 
 } catch (Exception $e) {
     $Logs = new Logs();
